* Renamed interface Observable to .hpp This reflects the coding-style of c++
* Introduced interface::DrawablePanel
This allows, for different types of panels, in rotation in dashboard. At
the same time, its serves as a basepanel, for SFML abstracts.

* Discovered a need for a ResourceFactory, for SFML uncoupling.
Found a need for a ResourceFactory, for media management. This would
allow gzip'd memory, and in-memory unpacking of rgb-values, etc.

* Introduced a prototype SmallImagePanel
This is the first of many small panels, that does different things,
depending on content and layout

* Discovered a need for panel inference
As a passive platform, 

* Fixed memory-leak in std::vector<interface::DrawablePanel*>
although its is frowned upon, to use a vector of ptrs, its necessary,
until the ResourceFactory is in use, otherwise the copy-constructor,
invoked in RAII would cause stack smashing. c++11 introduced the concept
of move-constructors, but its still not entirely c++11.

* Discovered a need for a ARM-target,
To build on arm, a seperate (and consistent) arm-target is needed.