* Introduced the concept of animators. In order to provide, sliding mechanism to panels efficiently, sfml provides a construct, view a floatrect, that controls a view.
Views, allow for multiple rectangle drawings to a backbuffer, and 
passing a sf::floatrect& down the graphical-update chain, allows for 
game mechanics to manipulate the view.

Its up to the initializer, to ensure the viewport and window dimensions
matches.

* Introduced the release build, 
the release-build uses a predefined size and fullscreen, 

* Ensured that progressbar uses Drawablepanel.
- Standardizing of names and stuff

The scene concept, currently uses a 'dashboard'. This uses, simple
mechanics to ping-pong expotentially back-and-forth, providing a gentle
viewing experience.